integration: gas_sponsorship: add in-kind tests (#295)
